+++ CHANGELOG.md
@@ -12,15 +12,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - check out how forgetting elixir does its popup
 - update/replace all text referring to memory boosters
 
+## WIP
+
+- improve admin console feedback on update
+
 ## [2.0.0]
 
 - add console command to view skill tracker
 - add dialog purchasing system for treatments
 - add player skill/perk tracking
 - add success/unnecessary/unaffordable responses
+- add support for crafting skill reset
 - remove deprecated buff
 - remove memory pills
-- sync progression data on save
 - update therapy prices on login and level-up
 - update to a21 mod-info file format
 - update to a21 references